Visualization/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pyzmq>=25.0.0
numpy>=1.21.0
PyYAML>=5.4.0

